ir/isa/pragma: using enum class instead of builtin uint32_t

+++ src/bootstrap/libnany/details/ir/isa/printer.inc.hpp
@@ -680,7 +680,7 @@ namespace // anonymous
 
 		void print(const Operand<Op::pragma>& operands)
 		{
-			if (operands.pragma < static_cast<uint32_t>(Pragma::max))
+			if (static_cast<uint32_t>(operands.pragma) < static_cast<uint32_t>(Pragma::max))
 			{
 				switch ((Pragma) operands.pragma)
 				{
